requirements.txt
click>=8.0.0
pyyaml>=6.0
jsonschema>=4.0.0
requests>=2.25.0
